Add Tree.searchItem(root, node)
The function searches the tree starting from root and compares by node.
It expects that there are no two exact nodes, which is impossible since
```TreeNode```s are compared by considering pointer address rather than
value.

The function is used in situation where you add a custom node, and then
there is a need to assign the added node to a pointer later-on. This
function helps in that situation.

When calling this funciton keep in mind the memory(stack) usage, since
it gets it's argument by value and the function is recursive.